requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
psutil==5.9.6
matplotlib==3.8.2
SQLAlchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
APScheduler==3.10.4
pillow==10.1.0

